Add files via upload
Millions of people visit Quora every month to seek answers for the questions they have. How amazing it would be for Quora if it knew which questions are similar. After all this could be used to instantly provide answers for questions which have already been answered. That's the task I solved.